Restructure CENNZnut project
