+++ src/test.rs
@@ -1,9 +1,28 @@
-// Copyright 2019 Centrality Investments Limited
+// Copyright (C) 2019-2020 Centrality Investments Limited
+// This file is part of CENNZnet.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+//!
+//! CENNZnut - Integration Tests
+//!
 
 #![warn(clippy::pedantic)]
 #![cfg(test)]
 
-use crate::{CENNZnutV0, Domain, Method, Module, Validate, ValidationErr, WILDCARD};
+use crate::method::Method;
+use crate::module::Module;
+use crate::{CENNZnutV0, Domain, Validate, ValidationErr, WILDCARD};
 use bit_reverse::ParallelReverse;
 use codec::{Decode, Encode};
 use pact::compiler::{Contract, DataTable};
